feat: 결과 페이지 metadata 분리 및 클라이언트 컴포넌트 구조 정리
